+++ .buildkite/scripts/run_models.sh
@@ -57,11 +57,11 @@ dbt compile --target "$db"
 dbt run --target "$db" --full-refresh
 dbt run --target "$db"
 dbt test --target "$db"
-dbt run --vars '{fivetran_platform__usage_pricing: true}' --target "$db" --full-refresh
-dbt run --vars '{fivetran_platform__usage_pricing: true}' --target "$db"
+dbt run --vars '{fivetran_platform__usage_pricing: true, fivetran_platform_using_connection: false, fivetran_platform_using_connector: true}' --target "$db" --full-refresh
+dbt run --vars '{fivetran_platform__usage_pricing: true, fivetran_platform_using_connection: false, fivetran_platform_using_connector: true}' --target "$db"
 dbt test --target "$db"
-dbt run --vars '{fivetran_platform__credits_pricing: false, fivetran_platform__usage_pricing: true}' --target "$db" --full-refresh
-dbt run --vars '{fivetran_platform__credits_pricing: false, fivetran_platform__usage_pricing: true}' --target "$db"
+dbt run --vars '{fivetran_platform__credits_pricing: false, fivetran_platform__usage_pricing: true, fivetran_platform_using_connector: true}' --target "$db" --full-refresh
+dbt run --vars '{fivetran_platform__credits_pricing: false, fivetran_platform__usage_pricing: true, fivetran_platform_using_connector: true}' --target "$db"
 dbt test --target "$db"
 dbt run --vars '{fivetran_platform__usage_pricing: false, fivetran_platform_using_destination_membership: false, fivetran_platform_using_user: false}' --target "$db" --full-refresh
 dbt run --vars '{fivetran_platform__usage_pricing: false, fivetran_platform_using_destination_membership: false, fivetran_platform_using_user: false}' --target "$db"