+++ integration_tests/tests/consistency/consistency__connector_daily_events.sql
@@ -7,34 +7,34 @@
 with prod as (
     select
         date_day,
-        connector_id, 
+        connection_id, 
         destination_id,
         count(*) as total_records
-    from {{ target.schema }}_fivetran_platform_prod.fivetran_platform__connector_daily_events
+    from {{ target.schema }}_fivetran_platform_prod.fivetran_platform__connection_daily_events
     group by 1, 2, 3
 ),
 
 dev as (
     select
         date_day,
-        connector_id, 
+        connection_id, 
         destination_id,
         count(*) as total_records
-    from {{ target.schema }}_fivetran_platform_dev.fivetran_platform__connector_daily_events
+    from {{ target.schema }}_fivetran_platform_dev.fivetran_platform__connection_daily_events
     group by 1, 2, 3
 ),
 
 final as (
     select 
         prod.date_day,
-        prod.connector_id,
+        prod.connection_id,
         prod.destination_id,
         prod.total_records as prod_total,
         dev.total_records as dev_total
     from prod
     left join dev 
         on dev.date_day = prod.date_day
-            and dev.connector_id = prod.connector_id
+            and dev.connection_id = prod.connection_id
             and dev.destination_id = prod.destination_id
 )