+++ macros/get_incremental_mar_columns.sql
@@ -2,6 +2,8 @@
 
 {% set columns = [
     {"name": "_fivetran_synced", "datatype": dbt.type_timestamp()},
+    {"name": "connection_id", "datatype": dbt.type_string()},
+    {"name": "connection_name", "datatype": dbt.type_string()},
     {"name": "connector_id", "datatype": dbt.type_string()},
     {"name": "connector_name", "datatype": dbt.type_string()},
     {"name": "destination_id", "datatype": dbt.type_string()},