+++ models/staging/tmp/stg_fivetran_platform__connection_tmp.sql
@@ -1,0 +1,4 @@
+{{ config(enabled=var('fivetran_platform_using_connection', True)) -}}
+
+select *
+from {{ var('connection') }}